fix: use specification on consult of order by id

+++ checkout/src/main/java/br/com/fiap/postech/adjt/checkout/model/dto/response/OrderCheckoutsResponse.java
@@ -7,7 +7,7 @@ public record OrderCheckoutsResponse(
         UUID orderId,
         List<CartItemResponse> items,
         String paymentType,
-        double value,
+        int value,
         String paymentStatus
 ) {
 }
